# user-info-svc/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "User Info API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/profile/me": {
      "get": {
        "tags": [
          "profile-controller"
        ],
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profile-controller"
        ],
        "operationId": "updateMe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "personalEmail": {
            "type": "string"
          },
          "maiEmail": {
            "type": "string"
          },
          "institute": {
            "type": "string"
          },
          "studentGroup": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "clothingSize": {
            "type": "string"
          },
          "social": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "contactEmail": {
            "type": "string"
          }
        }
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "login": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "personalEmail": {
            "type": "string"
          },
          "maiEmail": {
            "type": "string"
          },
          "institute": {
            "type": "string"
          },
          "studentGroup": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "clothingSize": {
            "type": "string"
          },
          "social": {
            "$ref": "#/components/schemas/SocialNetworks"
          },
          "contactEmail": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "SocialNetworks": {
        "type": "object",
        "properties": {
          "telegram": {
            "type": "string"
          },
          "vk": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "required": [
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "personalEmail": {
            "type": "string"
          },
          "maiEmail": {
            "type": "string"
          },
          "institute": {
            "type": "string"
          },
          "studentGroup": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "clothingSize": {
            "type": "string"
          },
          "social": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "contactEmail": {
            "type": "string"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "RefreshRequest": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}